src/cluster-config-api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pyyaml>=6.0
gunicorn>=21.0.0
